refactor(ui): extract table filters and delegates to ui.components
