+++ src/main.py
@@ -79,82 +79,6 @@ from core.services.ipc_service import (
 )
 
 
-class EmptyAreaClickFilter(QObject):
-    def __init__(self, table, parent=None):
-        super().__init__(parent)
-        self.table = table
-        self.rubber_band = QRubberBand(QRubberBand.Shape.Rectangle, self.table.viewport())
-        self.origin = QPoint()
-
-    def eventFilter(self, obj, event):
-        if obj != self.table.viewport():
-            return super().eventFilter(obj, event)
-
-        if event.type() == QEvent.Type.MouseButtonPress:
-            if event.button() == Qt.MouseButton.LeftButton:
-                item = self.table.itemAt(event.pos())
-                if not item:
-                    # Starting selection from empty area
-                    if not (event.modifiers() & (Qt.KeyboardModifier.ControlModifier | Qt.KeyboardModifier.ShiftModifier)):
-                        self.table.clearSelection()
-                    self.table.setCurrentItem(None)
-                    self.origin = event.pos()
-                    self.rubber_band.setGeometry(QRect(self.origin, QSize()))
-                    self.rubber_band.show()
-                    return True
-        elif event.type() == QEvent.Type.MouseMove:
-            if self.rubber_band.isVisible():
-                self.rubber_band.setGeometry(QRect(self.origin, event.pos()).normalized())
-                self.update_selection(event.modifiers())
-                return True
-        elif event.type() == QEvent.Type.MouseButtonRelease:
-            if self.rubber_band.isVisible():
-                self.rubber_band.hide()
-                return True
-        return super().eventFilter(obj, event)
-
-    def update_selection(self, modifiers):
-        rect = self.rubber_band.geometry()
-        selection_model = self.table.selectionModel()
-        
-        # Determine selection command
-        if modifiers & Qt.KeyboardModifier.ControlModifier:
-            command = QItemSelectionModel.SelectionFlag.Select
-        else:
-            command = QItemSelectionModel.SelectionFlag.ClearAndSelect
-            
-        command |= QItemSelectionModel.SelectionFlag.Rows
-        
-        selection = QItemSelection()
-        any_selected = False
-        
-        for row in range(self.table.rowCount()):
-            row_y = self.table.rowViewportPosition(row)
-            row_height = self.table.rowHeight(row)
-            # Selection happens if the rubber band vertically overlaps the row
-            if rect.bottom() >= row_y and rect.top() <= (row_y + row_height):
-                index = self.table.model().index(row, 0)
-                selection.select(index, index)
-                any_selected = True
-        
-        if any_selected:
-            selection_model.select(selection, command)
-        elif not (modifiers & Qt.KeyboardModifier.ControlModifier):
-            # If nothing touched and no Ctrl, clear everything
-            self.table.clearSelection()
-
-# --- HELPER FOR SORTING ---
-class SortableTableWidgetItem(QTableWidgetItem):
-    def __lt__(self, other):
-        v1 = self.data(Qt.ItemDataRole.UserRole)
-        v2 = other.data(Qt.ItemDataRole.UserRole)
-        if v1 is not None and v2 is not None:
-            try:
-                return float(v1) < float(v2)
-            except:
-                pass
-        return self.text() < other.text()
-
 from core.services.theme_service import (
     ACCENT_COLORS,
     CURRENT_ICON_THEME,
@@ -181,128 +105,12 @@ from core.services.theme_service import (
     parse_time_to_sec,
 )
 
-
-class SidebarItemDelegate(QStyledItemDelegate):
-    """
-    Delegate for the left panel category tree that guarantees the selected and hovered item's
-    icon and text render in pure black (#000000) over the accent highlight in both light and dark modes.
-    Formats non-selectable section headers with small font and muted text color.
-    """
-    def initStyleOption(self, option, index):
-        super().initStyleOption(option, index)
-        if index.data(Qt.ItemDataRole.UserRole) == "header":
-            option.state &= ~QStyle.StateFlag.State_MouseOver
-            option.state &= ~QStyle.StateFlag.State_Selected
-            option.font.setPointSize(8)
-            option.font.setBold(True)
-            app = QApplication.instance()
-            if app:
-                pal = app.palette()
-                placeholder_color = pal.color(QPalette.ColorRole.PlaceholderText)
-                option.palette.setColor(QPalette.ColorRole.Text, placeholder_color)
-                option.palette.setColor(QPalette.ColorRole.WindowText, placeholder_color)
-
-    def sizeHint(self, option, index):
-        size = super().sizeHint(option, index)
-        if index.data(Qt.ItemDataRole.UserRole) == "header":
-            return QSize(size.width(), 20)
-        return QSize(size.width(), 26)
-
-    def paint(self, painter, option, index):
-        if index.data(Qt.ItemDataRole.UserRole) == "header":
-            super().paint(painter, option, index)
-            return
-
-        is_selected = bool(option.state & QStyle.StateFlag.State_Selected)
-        is_hovered = bool(option.state & QStyle.StateFlag.State_MouseOver)
-
-        if is_selected or is_hovered:
-            opt = QStyleOptionViewItem(option)
-            self.initStyleOption(opt, index)
-
-            # Left panel selected/hovered icon MUST be pure black in both dark mode and light mode
-            icon = index.data(Qt.ItemDataRole.DecorationRole)
-            if isinstance(icon, QIcon) and not icon.isNull():
-                dec_size = opt.decorationSize if opt.decorationSize.isValid() and not opt.decorationSize.isEmpty() else QSize(18, 18)
-                src_pm = icon.pixmap(dec_size, QIcon.Mode.Normal)
-                if not src_pm.isNull() and src_pm.width() > 0 and src_pm.height() > 0:
-                    img = src_pm.toImage().convertToFormat(QImage.Format.Format_ARGB32)
-                    for y in range(img.height()):
-                        for x in range(img.width()):
-                            c = img.pixelColor(x, y)
-                            if c.alpha() > 0:
-                                img.setPixelColor(x, y, QColor(0, 0, 0, c.alpha()))
-                    black_pm = QPixmap.fromImage(img)
-                    black_ic = QIcon()
-                    black_ic.addPixmap(black_pm, QIcon.Mode.Normal)
-                    black_ic.addPixmap(black_pm, QIcon.Mode.Selected)
-                    black_ic.addPixmap(black_pm, QIcon.Mode.Active)
-                    opt.icon = black_ic
-
-            opt.palette.setColor(QPalette.ColorRole.Text, QColor("#000000"))
-            opt.palette.setColor(QPalette.ColorRole.HighlightedText, QColor("#000000"))
-            opt.palette.setColor(QPalette.ColorRole.WindowText, QColor("#000000"))
-
-            widget = option.widget
-            style = widget.style() if widget else QApplication.style()
-            style.drawControl(QStyle.ControlElement.CE_ItemViewItem, opt, painter, widget)
-        else:
-            super().paint(painter, option, index)
-
-
-class ToolbarHoverFilter(QObject):
-    """
-    Event filter applied to toolbar buttons to provide bold icon glow
-    and distinct hover/selection styling in both light and dark themes.
-    """
-    def __init__(self, main_window):
-        super().__init__(main_window)
-        self.main_window = main_window
-        self._action_icon_map = {
-            "action_add_url": "add_url",
-            "action_resume": "resume",
-            "action_stop": "stop",
-            "action_stop_all": "stop_all",
-            "action_delete": "delete",
-            "action_clear": "clear_completed",
-            "action_scheduler": "scheduler",
-            "action_options": "options",
-            "action_media_downloader": "media_downloader",
-        }
-        self._glow_icons = {}
-
-    def get_glow_icon(self, icon_name: str) -> QIcon:
-        if icon_name not in self._glow_icons:
-            from ui.icons import get_monochrome_icon
-            is_dark = self.main_window.is_dark_theme() if hasattr(self.main_window, "is_dark_theme") else True
-            stroke_color = QColor("#ffffff") if is_dark else QColor("#232629")
-            self._glow_icons[icon_name] = get_monochrome_icon(
-                icon_name,
-                color=stroke_color,
-                selected_color=stroke_color,
-                active_color=stroke_color,
-                glow=True
-            )
-        return self._glow_icons[icon_name]
-
-    def clear_cache(self):
-        self._glow_icons.clear()
-
-    def eventFilter(self, obj, event):
-        if isinstance(obj, QToolButton) and obj.isEnabled():
-            if event.type() in (QEvent.Type.Enter, QEvent.Type.MouseButtonPress):
-                action = obj.defaultAction()
-                if action:
-                    for attr, icon_name in self._action_icon_map.items():
-                        if getattr(self.main_window, attr, None) is action:
-                            obj.setIcon(self.get_glow_icon(icon_name))
-                            break
-            elif event.type() in (QEvent.Type.Leave, QEvent.Type.MouseButtonRelease):
-                if not obj.isDown() and not obj.underMouse():
-                    action = obj.defaultAction()
-                    if action:
-                        obj.setIcon(action.icon())
-        return super().eventFilter(obj, event)
+from ui.components import (
+    SortableTableWidgetItem,
+    EmptyAreaClickFilter,
+    SidebarItemDelegate,
+    ToolbarHoverFilter,
+)
 
 # --- CUSTOM DIALOG FOR DELETING COMPLETED ITEMS ---
 class MainWindow(QMainWindow):
